remote: Add --dry-run modes to download, upload, and delete commands
This mode, as is broader convention, goes through the motions of doing
the thing, as much as possible, but doesn't _actually_ do the thing.
That is, no changes should occur when --dry-run is active.

This is particularly useful for seeing what will happen if you're unsure
of how a path or argument is handled.  Dry runs of the list (ls) command
don't make sense and aren't included.

Resolves <https://github.com/nextstrain/cli/issues/149>.