remote.nextstrain_dot_org: Yield resource URL before deletion not after
A pre-action yield matches the messaging of the `remote delete` command
as of "remote delete: Update messaging to match pre-deletion behaviour"
(bb4ab4c) and the behaviour of the S3 remote as of "remote.s3: Rework
delete() to have a similar API to the other commands" (cc86a1a).

The post-action yield was an oversight and bug in the initial remote
implementation.