remote download: Use more of the resource path in the local filename
Includes all of the resource path except for the leading namespace part
(if any), e.g. /groups/X or /staging.

This reduces the potential for ambiguous filenames and makes it easier
to copy datasets/narratives between destinations (e.g. from one group to
another) while retaining the same path.  It is, however, a
**potentially-breaking change** if you're relying on the filenames of
the downloaded datasets/narratives (e.g. for automation).