update: Only update a specific runtime instead of all of them at once
With no arguments, the default runtime is updated.  The name of another
runtime to update instead may be provided as an argument.

In practice this isn't much of a behaviour change because only one
runtime currently supports updating (Docker); the others (native, AWS
Batch) just pass.  Existing users are unlikely to notice the change
unless they use multiple runtimes and Docker is not their default.  In
that case, `update` may stop updating Docker for them when it would have
done so previously.  The output of `update` should make this clear
though.

This change also improves the help, output, and other messaging of
`update` to not assume that Docker is the only runner in use.  (It
predated runners other than Docker in this codebase.)  As part of this,
the signature of RunnerModule.test_setup() now allows a "not supported"
(None) return value, in addition to the existing boolean for
success/failure.

Resolves <https://github.com/nextstrain/cli/issues/87>.