setup: Add command for automated set up of runtimes
Establishes a basic framework for implementations, even though none of
the runtimes support this currently.

However, it seems possible for the Docker runtime to implement, and more
importantly, it _will_ be a key piece of a new, forthcoming managed
Conda runtime.

For runtimes that don't support automated set up, this command will
still check for a manual set up and can optionally set the runtime as
the default if support is detected.  This behaviour overlaps some with
`check-setup`, but its presence can help with consistency of runtime
installation instructions.

Despite the command, like many of our commands, being a verb (to set up)
and not a noun (a setup), I still couldn't bring myself to name the
command "set-up".  It's ok.