Add a new runner for a fully-managed Conda runtime
This uses the Conda/Mamba ecosystem for the runtime environment, but the
user doesn't ever have to manage Conda themselves, thus potentially
making it the easiest install option when paired with the standalone
installer.  The Docker runtime is still more isolated and robust, but I
think this provides a really nice middle ground between "native" and
Docker.