shell: Customize the prompt ourselves instead of relying on the runtime's shell config
This removes a bit of tight coupling between `nextstrain shell` and the
up-until-now only runtime it supported: Docker.  We do this by arranging
for our own bashrc to be sourced at shell startup.  This change makes
way for supporting additional runtimes while retaining features like the
custom prompt to indicate to the user that they're in a `nextstrain
shell`.

Related-to: <https://github.com/nextstrain/docker-base/pull/84>