Prefix output of --dry-run modes with an indication of such
Nice to be crystal clear that what's said to be happening is not
_actually_ happening and make this consistent across commands with
dry run modes.

Providing this indication as an enterable context that doesn't require
changing subsequently-called code means that code doesn't have to be
muddied with conditional printing of indications.  While this wasn't
an issue for the existing dry run indications formerly used by the
`nextstrain remote` family of commands, it was a greater issue when
considering indications for the `nextstrain setup` command.  Thus this
work was motivated.

I used wrapt.ObjectProxy to implement LinePrefixer because a) it'll do a
*much* more thorough job of proxying than a basic __getattr__ override
and b) wrapt is already in our dep chain anyway. \o/