check-setup: Test that Snakemake, Augur, and Auspice are not just found but runnable
This catches situations where an executable is installed but doesn't
actually work, which, for example, we observed in testing of the managed
Conda runtime when a space was in the prefix path.  This could also
happen other ways, and it's easy enough for us to verify these programs
actually run (at least superficially enough for --version to work).