Rename "managed Conda" runtime to just "Conda"
A single word is simpler and better parallels the name of the Docker
runtime, which is a) named after the technology it uses (the "engine")
and b) also "managed" but doesn't say so in its name.

I'd used "managed Conda" because "Conda" by itself seemed too confusing:
we've long recommended that folks use Conda to manually setup an
environment suitable for the "native" runner.  There's a strong
historical association there that we'd be tangling up users in.  The
simpler "Conda" name does feel like the better name in the long run,
though.  Our hope is that we can avoid a lot of this confusion by
renaming "native" → "ambient" (more accurate anyhow), rephrasing our
install docs, and adding an FAQ entry or two.  We'll undoubtedly still
deal with some.

A push for this renaming and votes of confidence were given by
@victorlin and @huddlej in Slack.¹

¹ <https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1664232205090739?thread_ts=1662072691.294419&cid=C01LCTT7JNN>