runner.conda: Set HOME to our runtime data root when executing Micromamba
This avoids touching the ~/.mamba/ directory in the user's real homedir,
which is more than cosmetically important since installed files in
user-controlled Conda environments may be linked to the package cache
there.  We want to be able to invoke `micromamba clean` safely.

Note that HOME isn't set when running commands inside this runtime, only
during management operations with Micromamba.