setup: Handle the runtime argument using runner_module_argument()
…instead of handling it in a one-off way.  The initial reason for doing
something else here was so "choices" could be used, but that reason is
moot now that I've included the list of valid names in the error
messages produced by runner_module_argument().  It's good for usability
to maintain consistency across commands in handling of arguments.

This consistency also means `setup` will now accept unnormalized names
like "AWS-Batch" and future aliases.

Most changes in the diff are simple "runner" → "opts.runner" swaps due
to the type change of the latter meaning the former is no longer
necessary.