Rename "native" runtime/runner to "ambient"
This name better reflects what it is and has been a long time coming.
"Native" was never a name I was happy with, even when I introduced it.
The more descriptive name will help us explain what it is in
documentation and also further distinguish it from the new Conda
runtime, which is also "native" in the binary executable sense.

Continue to accept --native and "native" for backwards compatibility,
however.  The --native option continues to work anywhere it used to
previously but is hidden from --help output to discourage new use.
Anywhere runner names are accepted, e.g. in config as core.runner or in
command-line arguments to `check-setup`, "native" is accepted as well.