runner.conda: Run Micromamba in greater isolation from externally-activated environments
Filter out all CONDA_* and MAMBA_* env vars and set the root prefix and
no rc option for descendant processes of our micromamba invocation.

The CONDA_* vars are set during Conda environment activation and can
impact the operation and behaviour of Micromamba.  We don't need or want
that impact, as we don't want to make use of any of the activation
features of Conda/Mamba/Micromamba.