runner.conda: Pin the version of Micromamba used
For the same reason we pin deps with nextstrain-base, this ensures that
`nextstrain setup conda` can't break between one day and the next
because of a new Micromamba release.  This happened during my
development of the Conda runtime with the release of Micromamba 0.26.0¹,
and so has been on my mind to address.

The version used is overridable by an environment variable for debugging
and troubleshooting purposes, but it also serves as an escape hatch for
users stuck on older Nextstrain CLI versions who might need a newer
Micromamba for whatever reason.

We should generally try to keep the pinned version current as long as
new releases work ok.  It's too bad Dependabot doesn't support the Conda
ecosystem.²  If we're bad at remembering to update this over time, we
can setup a GitHub Action workflow to check for new releases and remind
us.

¹ https://github.com/mamba-org/mamba/issues/1979
² https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/about-dependabot-version-updates#supported-repositories-and-ecosystems