remote.nextstrain_dot_org: Embed images into narratives on upload
This permits local authoring of narratives that reference images on the
local filesystem without manual conversion (to remote images or embedded
images) before upload.

Unfortunately, both pyright and mypy (to a lesser degree) have enough
issues type checking the pyparsing API (and thus our usage of it), that
it is more tenable to skip their type checking of the new
nextstrain.cli.markdown module than to work around all their issues.
Hopefully this will be better in the future!  It's possible some of the
issues lay with the type annotations in pyparsing itself, and thus are
more easily improved.

A small number of new tests help ensure the parsing and embedding of
images works as expected, though more should be added with time as we
run across additional real-world test cases.  The framework established
here makes that as quick as adding another Markdown file to the test
data.