dev: Avoid PyOxidizer 0.23.0 in CI
Builds with this version are seemingly-successful until runtime when
this error is thrown:

    Traceback (most recent call last):
      …
      File "cryptography.hazmat.primitives.asymmetric.ec", line 11, in <module>
      File "cryptography.hazmat._oid", line 7, in <module>
    ImportError: …/build/x86_64-unknown-linux-gnu/debug/installation/lib/cryptography/hazmat/bindings/_rust.abi3.so: undefined symbol: PyExc_BaseException

Why 0.23.0 results in a bad build remains unclear, though I'll continue
to try to track down the cause.  I tested and confirmed it's not the
Python 3.10.4 → 3.10.8 upgrade noted in the 0.23.0 release notes¹.
Maybe it's the SETUPTOOLS_USE_DISTUTILS=stdlib change?

For now, avoiding 0.23.0 means we'll use the previous release, 0.22.0,
which still works well in my local testing and is what we built our last
release, 5.0.1, with.  I'm hopeful the next release of PyOxidizer will
fix the issue; the single version exclusion will help us notice if it
doesn't.

This issue was first identified in CI runs for an unrelated PR.²

¹ https://github.com/indygreg/PyOxidizer/releases/tag/pyoxidizer%2F0.23.0
² https://github.com/nextstrain/cli/pull/235