view: Accept paths to a specific dataset or narrative file
The data directories used by Auspice will be inferred from the file's
path and the given dataset or narrative will be opened in a browser by
default (instead of Auspice's dataset/narrative listing).

This makes the UX for viewing a specific dataset or narrative much
nicer, with fewer steps and a more direct translation of intent into
invocation (e.g. "I want to view this dataset" vs. "I want to view this
dataset, so I have to specify its parent directory").