remote.nextstrain_dot_org: Add env var to disable narrative image embedding
It's often useful to have an escape hatch, especially for things in the
critical path (here, the upload path).  If for any reason some narrative
file finds a bug in the image embedding code and causes a crash or
corrupts the Markdown, we'll appreciate having a way to disable it at
runtime.