standalone-installer-unix: Check we can exec the installed binary…
…and error if we can't.  Previously, failures to exec the installed
binary would be ignored and the installer would continue to suggest the
final shell setup steps (which would result in errors when run).

This catches an observed issue when installing our x86_64 binary on an
arm64 Mac before Rosetta 2 is enabled.

    bash: line 72: /Users/blab/.nextstrain/cli-standalone/nextstrain: Bad CPU type in executable
    ______________________________________________________________________________

    Nextstrain CLI () installed to /Users/blab/.nextstrain/cli-standalone.

Stopping early is better than suggesting bad next steps bound to fail.

Related-to: https://github.com/nextstrain/docs.nextstrain.org/pull/147