Support even older glibc (and thus Linux distro) versions for our standalone binary
Arranges for `pyoxidizer build` to run inside of a manylinux2014
container instead of on the ambient host, which yields the following
glibc minimum version improvements compared to our existing binaries
produced in CI (as reported by `pyoxidizer analyze`):

    Minimum Version: 2.27 → 2.17
    Minimum Distro Versions:
      Debian 10     → 8
      Fedora 28     → 19
      OpenSUSE 15.3 → 12.3
      RHEL 8        → 7
      Ubuntu 18.04  → 14.04

The manylinux containers are readily available, well-maintained, and
designed exactly for this purpose.  The standalone Python distributions
used by PyOxidizer are themselves limited to glibc ≥2.16, so we're now
basically at that limit!  manylinux2010 versions are too old.

Related-to: https://github.com/nextstrain/cli/issues/234