Add a new runner for a Singularity runtime based on the Docker runtime
It's been my intention to support Singularity since day 1, but I never
got around to implementing it because usage seemed little to
non-existent amongst Nextstrain users I knew about.  (Though maybe
that's a chicken-or-the-egg problem!)  We've since had users report
their ad-hoc use of Singularity with Nextstrain, however, and
Singularity is often the only containerization method supported by HPC
systems.  So it seems important to support.

Resolves <https://github.com/nextstrain/cli/issues/2>.