Standardize use of "runtime" and related terminology
Use "runtime" ~exclusively in user-facing output and documentation,
replacing other terms such as "computing environment", "build
environment", and "runner".

"Computing environment" and "ambient environment" are now used only
sparingly and in the general sense.  "Computing platform" is now used to
describe the things on which our runtimes are based (Docker, Conda,
etc).

The code still makes a distinction between "runners" and "runtimes"
that's, albeit, fuzzy in places, but nevertheless useful.  The runtime
is the environment itself (e.g. a container image or a Conda
environment), while the runner is the code (Python module) which
arranges for Nextstrain CLI commands to execute things inside the
runtime.  Runners and runtimes have a 1:1 mapping.