runner.conda: Hint in check-setup output if `setup conda` needs to be run
The Conda runner is the first that can be setup all by itself, as long
as the OS is supported.  The current design of check-setup doesn't
really account for the separate state of supported vs. set up, so I
conditionalize what we report from test_setup().

As we add more automatic setup support for runtimes¹, it will probably
be good to change check-setup to consider support status and set up
status separately.

¹ e.g. <https://github.com/nextstrain/cli/issues/220>