setup-miniconda: Use Mambaforge
The option to use mamba-version in this action is still marked as
experimental and has shown to be problematic with transient failures.

An alternative is to use Mambaforge, which is a Miniforge variant that
comes with Mamba in the base environment. It uses the conda-forge
channel as the only default channel¹.

However, conda-forge must be added to the channels input to the
setup-miniconda action, since it creates and uses a separate condarc
file which contains the specified channels along with defaults², and
defaults should be avoided at least for the Python version³.

¹ https://github.com/conda-forge/miniforge/blob/871e98647a95b99a769dd9f40669d06c44514bca/README.md
² https://github.com/nextstrain/cli/pull/249#issuecomment-1382433970
³ https://github.com/nextstrain/cli/pull/249#issuecomment-1381300485