docs: Add a glossary, primarily for development
Documents the meanings I was working from in «Standardize use of
"runtime" and related terminology» (8a1e87b).  Good to make explicit
terms that have been largely implicit until now (with the exception of
"runtime").

I tried to be precise and non-overlapping to make the relationships
between these terms sensible, but nevertheless they may want
clarifications or revisions later.  In any case, the bigger point of the
glossary is to give a broad sense of the terms, rather than definitions
that would satisfy a stickler.