standalone-installer-unix: Check for Rosetta 2 in a way that works on macOS 11 too
The previous check worked only on macOS 12 and 13, because the
executable path on macOS 11 is:

    /Library/Apple/usr/libexec/oah/oahd

On all macOS versions, however, the process runs as the _oahd user.  In
case the _oahd user doesn't exist, direct stderr to the bit bucket to
avoid a noisy error being confusing to the person running the installer.

Troubleshooting happened on Slack¹ with the help of @cassiawag and
@corneliusroemer.

¹ <https://bedfordlab.slack.com/archives/C0K3GS3J8/p1675714639052449>