build: Add --no-logs option for AWS Batch builds
It can be annoying to sit thru the fetching and printing of large reams
of log lines when attaching to a completed AWS Batch build to download
the results.  This is particularly true when downloading small bits of
the results via --download, as the log fetching often takes longer than
the selective download does!