remote.nextstrain_dot_org: Fix uninterpolated placeholder in error message
When an unknown file was given for upload, the error message incorrectly
included the literal string "{files}" instead of replacing it with the
actual list of files.

Although this string doesn't need to be an f-string, the other similar
error messages below it do, and preserving the parallel construction of
the error messages is helpful for readability.