runner.aws_batch: Fix error on job submission with botocore ≥1.28.0
Job submissions which triggered the auto-registration of a new job
definition (i.e. because a new image was needed) were failing with
botocore ≥1.28.0 (released 24 Oct 2022), because a new field that is
invalid for registering new job definitions started being returned with
the existing job definition used as a base.  This resulted in errors
from `nextstrain build` like:

    Submitting job
    Parameter validation failed:
    Unknown parameter in input: "containerOrchestrationType", must be one of: jobDefinitionName, type, parameters, schedulingPriority, containerProperties, nodeProperties, retryStrategy, propagateTags, timeout, tags, platformCapabilities, eksProperties
    Job submission failed!

So far this bug only manifests when installing Nextstrain CLI from the
Bioconda package (and potentially only with Mamba's dependency solver)
because of interactions between s3fs, botocore, and aiobotocore.  With
the Bioconda package (and Mamba), the dependency solution uses a new
botocore and old s3fs, which doesn't use aiobotocore and thus doesn't pin
an older botocore.¹  An old s3fs is allowed because the Bioconda package
doesn't specify a lower bound.

In contrast, our Python package installed by Pip _does_ specify a lower
bound for s3fs², which then requires aiobotocore, which then requires an
older botocore where we don't encounter this bug.

A better approach than this hardcoded list of fields will come in the
following commit.

¹ aiobotocore is a giant monkeypatch of botocore, and thus has very
  narrow version bounds on botocore.

² See <https://github.com/nextstrain/cli/blob/e9f11e60/setup.py#L100-L123>
  for why.

Related-to: <https://github.com/nextstrain/cli/pull/108>
Related-to: <https://github.com/nextstrain/cli/pull/134>