standalone-installer-unix: Use `eval` instead of `source` to avoid issues on old versions of Bash
In Bash prior to 4.0 (released in 2009), the `source` builtin does not
properly read input from a pipe, named or otherwise.  When given a pipe,
such as via process substitution (i.e. `<(…)`), it silently does nothing
and exits 0.  This can cause a SIGPIPE in the process on other end of
the pipe and, in Python, a BrokenPipeError.

Unfortunately, macOS ships with Bash 3.2.57 (released in 2007) and, for
legal reasons, is unlikely to ever be upgraded.  Switching to `eval`
with command substitution (i.e. `$(…)`) avoids the issue for Bash users
on macOS and results in no functional change for everyone else.

We _will_ have to ensure that the output of `nextstrain init-shell`
remains below command line length limits since the output is now
interpolated into the command line, but I don't think this will ever be
an issue since the initialization sequence should remain pretty short.

Initial problem report and troubleshooting with @alliblk and @victorlin
happened in Slack.¹

¹ <https://bedfordlab.slack.com/archives/C7SDVPBLZ/p1677871789858519>