dev: Stop using Ubuntu 18.04 in CI; it will soon be dropped by GitHub Actions
Ubuntu 18.04 was to be dropped by GitHub on 1 Dec 2022, but they later
extended that to April 2023.¹  That's in a few weeks and I've already
noticed the brownouts, so time to get moving.

Building on 18.04 doesn't matter anymore thanks to the containerized
build environment using manylinux images, but it'd still be nice to test
on older versions.  We could use containers for this too, but doing so
is an exercise for the future.

¹ <https://github.blog/changelog/2022-08-09-github-actions-the-ubuntu-18-04-actions-runner-image-is-being-deprecated-and-will-be-removed-by-12-1-22/>