runner.conda: Explicitly specify a nextstrain-base version when updating
This helps Micromamba (or really, libsolv) take the action we want,
upgrading to the latest version, as otherwise it sometimes chooses a
solution that results in no upgrade.¹

It's also good to be explicit about the version we expect, and this
behaviour is closer to what we do when updating other runtimes as well.
The update messaging is also improved in the no-op case by having the
desired version on hand.

¹ e.g. <https://github.com/nextstrain/cli/pull/266#issuecomment-1470880832>