Add newline before eval command
Previously, the command could result in unexpected errors if the shell
init file did not end with a newline (for reasons beyond our control).

If the file already ends with a newline, this comes with the added
benefit of using whitespace to distinguish the "scope" of this command
as separate from whatever was there before.

Use printf to ensure '\n' is expanded in bash, zsh, and hopefully all
other shells.