runner.conda: Explicitly get latest `main` version
Learned through the conda-base repo's CI job that Anaconda's
`/release/{owner_login}/{package_name}/latest` API endpoint returns the
latest package version regardless of labels.¹ This means that users can
accidentally install test versions of the nextstrain-base package.

I tried various endpoints from the API docs² to see if we can pass the
label as a URL param that can allow the filtering to be done on server,
but there were no leads in the docs and nothing I tried worked.
So this commit adds an explicit check for the latest `main` version
by pulling all package files and then filtering for the label and
finding the max `version`.

If the check for the latest `main` version fails, then this will return
to the previous behavior of using `latest` with a warning that the
latest version will be used regardless of label.

¹ https://github.com/nextstrain/conda-base/issues/26
² https://api.anaconda.org/docs