runner.singularity: Support for build's --cpus and --memory options
Only available since Singularity 3.10.0 and on systems with cgroups v2.¹
With older Singularity versions, these options will continue to be
ignored (as before this change).  If the Singularity version is new
enough but the system doesn't support cgroups v2 (or it is
administratively disabled in the Singularity config), then Singularity
will error when these options are used.

¹ <https://docs.sylabs.io/guides/latest/user-guide/cgroups.html#requirements-linux-cgroups>