dev: Work around some mypy limitation when type checking a lambda
With the lambda syntax, mypy raises these errors for currently-to-me
inscrutable reasons:

    nextstrain/cli/runner/singularity.py:24: error: Name "docker" already defined (by an import)  [no-redef]
    nextstrain/cli/runner/singularity.py:136: error: Cannot determine type of "singularity_version_at_least"  [has-type]
    Found 2 errors in 1 file (checked 54 source files)

The errors make no sense: the "docker" "redefinition" is the sole
definition, the import, and it is not redefined anywhere in our code
(but maybe it is in mypy's mangling of the code?), and the type of
"singularity_version_at_least" seems clear to me (but maybe mypy
can't/isn't forward looking here?).

Why this workaround works is unknown to me, but I guessed it would
perturb the code path in mypy enough and, lo, it does and it's
unoffensive enough to use.