Revert "Require docutils <0.16 in dev deps to avoid a version conflict"
This reverts commit d2463b72c153ea6720d920baff01003d51ec4c04.

Pip's dependency resolution around docutils no longer appears to be an
issue.  We also no longer test on Python 3.6.1 in CI but instead use
3.6.15.

This pin was causing issues in CI tests itself since it was preventing
use of newer docutils with fixes for deprecated pkg_resource APIs.  The
older docutils versions caused warnings at test time.

Since docutils is also a non-dev dependency and unpinned there, it's
also nice to use the same version in dev/test as in non-dev installs.