view: Override HOST inside a Docker container with an alternative method
Passing an additional --env=HOST=0.0.0.0 arg to `docker run` is a hacky
implementation encapsulation break.  It only works to override the
HOST=127.0.0.1 value provided via extra_env because the Docker runner
translates extra_env to --env options that come before the additional
args.  The implementation of extra_env is about to change, however, and
the hacky --env=HOST=0.0.0.0 arg here would no longer override it.

All this would be simpler if the container shared the host's network
interfaces (i.e. --network=host) instead of using separate interfaces
and bridging to the host, but doing so assumes the container (i.e.
dockerd) is running on the same host as we are.  True most of the time,
but not guaranteed.  This is the more minimal change anyway.