devel: Replace `realpath` from GNU coreutils with `cd …; pwd`
For compat on systems that don't have GNU coreutils, like macOS by
default.  Motivated by noticing in macOS CI jobs that devel/pytest was
reporting "command not found: realpath" which results in `cd ""`, which
results in no change of directory.  But that's ok in the CI context so
tests still worked.

Anyhow, the usage in devel/pytest was copypasta (from me, two weeks
ago!) that never warranted realpath.