Declare minimum version requirement for pyparsing
We rely on the pyparsing 3.0.0 API.  It was released well before our
first use of it, but still good to cross Ts and dot Is because older
pyparsing might be used by other deps in an environment.  I had cause to
think about this when debugging just now, so fix it up.