markdown: Fix parsing of `![alt](url "title")` under pyparsing ≥3.1.0b1
The recent release of pyparsing fixed a bug in SkipTo¹ "that has been
there for a very very long time"² and which we were unintentionally
relying upon.  The fix caused our parsing of

    ![alt](url "title")

to break as SkipTo(White()) no longer matched due to our grammar
ignoring whitespace by default.  We didn't notice until 3.1.0 was
released a couple days ago.

Explicitly don't ignore whitespace for this grammar element.  This
change is backwards compatible with older versions of pyparsing.

Resolves <https://github.com/nextstrain/cli/issues/293>.

¹ <https://github.com/pyparsing/pyparsing/commit/9d789cbc>
² <https://github.com/pyparsing/pyparsing/issues/475>