Update link to mypy editor integrations
The section title was renamed so the previous link broke does not jump
to the section anymore. I considered using a permalink to protect from
future changes, but figured it's more likely that the list of editor
integrations will be updated with useful information than the section
title renamed again.