argparse: Don't visit aliased commands twice in walk_commands()
This avoids things like testing the --help output of both `nextstrain
remote list` and `nextstrain remote ls`.  The primary name is visited
first, so it's the aliases we'll skip.

Note that this only applies to simple aliases (i.e. using the "aliases"
feature of argparse), not more complex aliases like our `nextstrain
deploy`.