docs: Generate the command doc pages statically instead of dynamically
This ditches sphinx-argparse and the "argparse" directive in favor of
our own static generation of rST from the argument parsers.

The main motivation here is to emit the Sphinx "program" and "option"
directives in these pages so we can use the Sphinx "option" role (e.g.
:option:`--foo`) to cross-reference command options.  Cross-referencing
like this is very useful because it means readers don't have to chase
down the docs for an unfamiliar option themselves.

Secondary motivations include having more control over the formatting of
the pages and making more apparent the overall structure of the pages
(since we're both authoring the rST source and can see the generated
output, unlike with the "argparse" directive).

As an example of where more formatting control is useful: now we can put
the usage example at the very top, above the command description, like
it is in --help output.  The "argparse" directive put the usage below
the description, but this made descriptions confusing since they often
refer to placeholders in the usage.

The statically generated files are checked into version control for
visibility and because they should be the exact same for any maintainer.
To ensure the copies in version control stay up-to-date, the tests
assert that generated docs are up-to-date.  To ensure that the built
docs stay up-to-date even if version control diverges, the docs build
process automatically regenerates the files if needed.

Note that while generate-command-doc accesses private argparse
internals, this is precisely what sphinx-argparse does as well so we're
not any less insulated from breakage.  The private internals we use are
limited and, given historical development of argparse, seem unlikely to
change routinely.