help: Distinguish between `command x y` and ``a b c`` when rendering help strings
Use the former for examples of command invocations and the former for
generic literal text snippets.  In the text terminal context, this
leaves the generic literals unadorned (they're already readable as-is)
and the command invocations surrounded by backticks (so they stand out
as such from surrounding text).  In the web context, both forms are
distinguished from the surrounding text the same way.