docs: Add a "command-reference" role for referring to a command's help
This bit of rST:

    See :command-reference:`nextstrain remote` for more details.

now renders in the text terminal context as:

    See `nextstrain build --help` for more details.

and in the web context renders "nextstrain build" as a link to the
command's doc page.

There are fancier/better approaches—such as creating an extension and
using Sphinx.add_object_type()¹—but the approach here seemed like the
simplest way to address the different capabilities of the terminal vs.
web.  We may want to switch to a different approach in the future,
however.

¹ <https://www.sphinx-doc.org/en/master/extdev/appapi.html#sphinx.application.Sphinx.add_object_type>