dev: Don't force Python 3.6 in devel/setup-venv
This requirement was a well-intentioned attempt to try to avoid
introducing backwards compatibility issues that can arise if, e.g.,
developing on 3.10 and not realizing something won't work on 3.6.  But
it turns out to make testing locally on different versions of Python
less convenient since devel/setup-venv can't be used.  Rely on CI tests
to catch backwards/forwards compatibility issues instead.