check-setup: Document usage ourselves rather than rely on argparse's default
This is an already established pattern we use with other commands where
we want nicer usage specs than what argparse provides.  With this commit
we preserve the:

    [<runtime> [<runtime> ...]]

notation on all versions of argparse, regardless of Python version.
That's the form produced by argparse on Python <3.10, but later versions
would shorten it to:

    [<runtime> ...]

which is less descriptive in my eyes.

Like our other existing examples of custom usage, we take the
opportunity to break out -h from normal invocations and expand it to
--help for clarity.